getUserTimeline updated for Twitter API 1.1
also deleted the AIR.swc as it is not up to date. replaced with
Tweetr1.1.swc.  getUserTimeline() is the only method updated. other
methods will (probably) log in to the new API but you will need to
check variable name changes between API1 and the new 1.1 to allow them
to work as expected.